requirements.txt
numpy
openai
pandas
python-dotenv
seaborn
streamlit
langchain
langchain-community
langchain-experimental
